backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
openai==2.22.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
python-dotenv==1.0.1
eval-type-backport==0.3.1
